Start subprocesses without stealing focus
If you are using the computer while mining, it's really fustrating when
you are typing and a new mining window opens and steals focus.

This starts subprocesses minimized without becoming the active window
and without stealing focus.

Uses the windows native kernel32 call to start processes, instead
of Start-Process, since there is no way in powershell to open a window
without it becoming active.